Extract `cvd/fetch/host_tools_target.{cc,h}`
Bug: b/424021019
Test: bazel test '//...' --test_summary=terse
